♻️Refactor : 지도 상수 파일명 변경(home->map.ts)
